Move Filter and RootEntry to lib.rs
